Update Users model to include profile image and ypdate view templates to reflect this change
