+++ models/Recipe.js
@@ -11,7 +11,14 @@ const reviewSchema = new mongoose.Schema(
       min: 1,
       max: 5,
       default: 5
-    }
+    },
+    user: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
+      required: true
+    },
+    userName: String,
+    userAvatar: String
   }, {
   timestamps: true
 }
