+++ views/partials/header.ejs
@@ -20,7 +20,7 @@
 <body>
     <nav>
     <div class="navbarContainer">
-      <a id="logo" href="/recipes">RecipeTracker</a>
+      <a id="logo" href="/recipes">Allrecipes Was Taken</a>
       <ul>
         <% if(user) { %>
           <li>
